feat(app-router): Pages Router → App Router 전환 (M3)
- app/ 라우트 도입: 루트 layout(html/body/metadata/Providers) + (site) 그룹 layout(사이드바 쉘)
- /resume 는 (site) 그룹 밖 독립 문서로 유지(사이드바 없음)
- getStaticProps → 서버 컴포넌트 async 페치(projects, revalidate 1h)
- getStaticPaths → generateStaticParams(blog/[slug]), Next 15 params Promise await
- next/head → 메타데이터 API; 인터랙티브는 'use client' 리프로 분리
  (blogIndex 카테고리탭, blog/toc 스크롤스파이, blog/codeBlock 복사, resumeDoc 프린트+styled-jsx, sidebar/footer/darkModeToggleBtn)
- sidebar: next/router useRouter → next/navigation usePathname
- _app/_document/pages/ 및 미사용 api/hello 제거, components/layout.tsx 제거
- resume 폰트: React 19 link hoisting으로 next/head 대체

빌드 검증: next build 통과(11 라우트, generateStaticParams 4 slugs, projects ISR)
