+++ app/(site)/blog/page.tsx
@@ -1,0 +1,7 @@
+import BlogIndex from "../../../components/blogIndex"
+
+export const metadata = { title: "Blog" }
+
+export default function Blog() {
+  return <BlogIndex />
+}
