+++ app/resume/page.tsx
@@ -1,0 +1,10 @@
+import ResumeDoc from "../../components/resumeDoc"
+
+export const metadata = {
+  title: "이력서 · 백엔드 개발자",
+  description: "오세준(Sejune Oh) — C#/.NET 백엔드 개발자 정식 이력서",
+}
+
+export default function Resume() {
+  return <ResumeDoc />
+}
