+++ components/darkModeToggleBtn.tsx
@@ -1,3 +1,5 @@
+"use client"
+
 import { useTheme } from "next-themes"
 import { useEffect, useState } from "react"
