+++ components/footer.tsx
@@ -1,3 +1,5 @@
+"use client"
+
 import { useState } from "react"
 import type React from "react"
 import Link from "next/link"
