--- pages/404.tsx
@@ -1,5 +1,0 @@
-export default function PageNotFound() {
-  return <>
-    <h1>Page not found</h1>
-  </>
-}
